requirements.txt
grpclib
h2
hpack
hyperframe
multidict
numpy
protobuf
python-dateutil
pytz
six
typing_extensions
tzdata
viam-sdk
pyinstaller
google-api-python-client